# core/cmd/server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Top-Rated Web-based Linux Server Management Tool",
        "title": "1Panel",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/api/v2",
    "paths": {
        "/core/auth/api/generate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "generate api key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "x-panel-log": {
                    "BeforeFunctions": [],
                    "bodyKeys": [],
                    "formatEN": "generate api key",
                    "formatZH": "生成 API 接口密钥",
                    "paramKeys": []
                }
            }
        },
        "/core/auth/api/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update api config",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.ApiInterfaceConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "x-panel-log": {
                    "BeforeFunctions": [],
                    "bodyKeys": [
                        "ipWhiteList",
                        "apiTrustedProxies"
                    ],
                    "formatEN": "update api config =\u003e IP Allowlist: [ipWhiteList], API Trusted Proxies: [apiTrustedProxies]",
                    "formatZH": "更新 API 接口配置 =\u003e IP 白名单: [ipWhiteList], API 可信代理: [apiTrustedProxies]",
                    "paramKeys": []
                }
            }
        },
...
        "/core/auth/current": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Load current user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.CurrentUserInfo"
                        }
                    }
                }
            }
        },
        "/core/auth/current/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update current user info",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.CurrentUserUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/core/auth/expired/reset": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Reset system password expired",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.PasswordUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "x-panel-log": {
                    "BeforeFunctions": [],
                    "bodyKeys": [],
                    "formatEN": "reset an expired Password",
                    "formatZH": "重置过期密码",
                    "paramKeys": []
                }
            }
        },
        "/core/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全入口 base64 加密串",
                        "name": "EntranceCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.UserLoginInfo"
                        }
                    }
                }
            }
        },
        "/core/auth/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/core/auth/mfa": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "System Setting"
                ],
                "summary": "Load mfa info",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.MfaCredential"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_utils_mfa.Otp"
                        }
                    }
                }
            }
        },
        "/core/auth/mfa/bind": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "System Setting"
                ],
                "summary": "Bind mfa",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.MfaCredential"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "x-panel-log": {
                    "BeforeFunctions": [],
                    "bodyKeys": [],
                    "formatEN": "bind mfa",
                    "formatZH": "mfa 绑定",
                    "paramKeys": []
                }
            }
        },
        "/core/auth/mfa/close": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "System Setting"
                ],
                "summary": "Close mfa",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "x-panel-log": {
                    "BeforeFunctions": [],
                    "bodyKeys": [],
                    "formatEN": "close mfa",
                    "formatZH": "mfa 关闭",
                    "paramKeys": []
                }
            }
        },
        "/core/auth/mfalogin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login with mfa",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.MFALogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.UserLoginInfo"
                        },
                        "headers": {
                            "EntranceCode": {
                                "type": "string",
                                "description": "安全入口"
                            }
                        }
                    }
                }
            }
        },
...
        "/core/auth/passkey/del": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Delete passkey",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/core/auth/passkey/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "List passkeys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.PasskeyInfo"
                            }
                        }
                    }
                }
            }
        },
        "/core/auth/passkey/register/begin": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Begin passkey registration",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.PasskeyRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.PasskeyBeginResponse"
                        }
                    }
                }
            }
        },
        "/core/auth/passkey/register/finish": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Finish passkey registration",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/core/auth/welcome": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Timestamp": []
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get welcome page",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_1Panel-dev_1Panel_core_app_dto.ApiInterfaceConfig": {
            "type": "object",
            "properties": {
                "apiInterfaceStatus": {
                    "type": "string"
                },
                "apiKey": {
                    "type": "string"
                },
                "apiKeyValidityTime": {
                    "type": "integer"
                },
                "apiTrustedProxies": {
                    "type": "string"
                },
                "ipWhiteList": {
                    "type": "string"
                }
            }
        },
...
        "github_com_1Panel-dev_1Panel_core_app_dto.CurrentUserInfo": {
            "type": "object",
            "properties": {
                "apiInterfaceStatus": {
                    "type": "string"
                },
                "apiKey": {
                    "type": "string"
                },
                "apiKeyValidityTime": {
                    "type": "integer"
                },
                "apiTrustedProxies": {
                    "type": "string"
                },
                "authSource": {
                    "type": "string"
                },
                "authSourceStatus": {
                    "type": "string"
                },
                "complexitySetting": {
                    "type": "string"
                },
                "ipWhiteList": {
                    "type": "string"
                },
                "mfaInterval": {
                    "type": "integer"
                },
                "mfaStatus": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nodeRoles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_1Panel-dev_1Panel_core_app_dto.CurrentUserNodeRole"
                    }
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Custom Token Format, Format: md5('1panel' + API-Key + UnixTimestamp).\n```\neg:\ncurl -X GET \"http://{host}:{port}/api/v2/toolbox/device/base\" \\\n-H \"1Panel-Token: \u003c1panel_token\u003e\" \\\n-H \"1Panel-Timestamp: \u003ccurrent_unix_timestamp\u003e\"\n```\n- `1Panel-Token` is the key for the panel API Key.",
            "type": "apiKey",
            "name": "1Panel-Token",
            "in": "Header"
        },
        "Timestamp": {
            "description": "- `1Panel-Timestamp` is the Unix timestamp of the current time in seconds.",
            "type": "apiKey",
            "name": "1Panel-Timestamp",
            "in": "header"
        }
    }
}